refactor: messages package
- simplify naming
- move triplestoremessages to the store package (under messages)
