+++ webapi/src/main/scala/org/knora/webapi/messages/v1/responder/graphdatamessages/GraphDataMessagesV1.scala
@@ -22,11 +22,11 @@ package org.knora.webapi.messages.v1.responder.graphdatamessages
 
 import org.knora.webapi._
 import org.knora.webapi.messages.v1.responder.resourcemessages.{PropsV1, ResourceInfoV1}
-import org.knora.webapi.messages.v1.responder.{KnoraRequestV1, KnoraResponseV1}
 import org.knora.webapi.messages.v1.responder.usermessages.{UserDataV1, UserProfileV1}
+import org.knora.webapi.messages.v1.responder.{KnoraRequestV1, KnoraResponseV1}
 import spray.json._
 
-import scala.collection._
+import scala.collection.breakOut
 
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -104,6 +104,7 @@ case class GraphDataEdgeV1(label: Option[String], from: IRI, to: IRI)
 
 object GraphDataV1JsonProtocol extends DefaultJsonProtocol with NullOptions {
 
+    import org.knora.webapi.messages.v1.responder.resourcemessages.ResourceV1JsonProtocol._
     import org.knora.webapi.messages.v1.responder.usermessages.UserDataV1JsonProtocol._
 
     implicit object GraphDataV1JsonFormat extends JsonFormat[GraphV1] {
