+++ webapi/src/main/scala/org/knora/webapi/responders/v1/HierarchicalListsResponderV1.scala
@@ -24,10 +24,8 @@ import akka.actor.Status
 import akka.pattern._
 import org.knora.webapi._
 import org.knora.webapi.messages.v1.responder.listmessages._
-import org.knora.webapi.messages.v1respondermessages.listmessages._
-import org.knora.webapi.messages.v1respondermessages.triplestoremessages.VariableResultsRow
-import org.knora.webapi.messages.v1respondermessages.usermessages.UserProfileV1
-import org.knora.webapi.messages.v1storemessages.triplestoremessages.{SparqlSelectRequest, SparqlSelectResponse, VariableResultsRow}
+import org.knora.webapi.messages.v1.responder.usermessages.UserProfileV1
+import org.knora.webapi.messages.v1.store.triplestoremessages.{SparqlSelectRequest, SparqlSelectResponse, VariableResultsRow}
 import org.knora.webapi.util.ActorUtil._
 
 import scala.annotation.tailrec
