+++ webapi/src/main/scala/org/knora/webapi/responders/v1/SearchResponderV1.scala
@@ -26,10 +26,7 @@ import org.knora.webapi._
 import org.knora.webapi.messages.v1.responder.ontologymessages.{EntityInfoGetRequestV1, EntityInfoGetResponseV1, ResourceEntityInfoV1}
 import org.knora.webapi.messages.v1.responder.searchmessages._
 import org.knora.webapi.messages.v1.responder.valuemessages.KnoraCalendarV1
-import org.knora.webapi.messages.v1respondermessages.ontologymessages.ResourceEntityInfoV1
-import org.knora.webapi.messages.v1respondermessages.searchmessages._
-import org.knora.webapi.messages.v1respondermessages.triplestoremessages.SparqlSelectResponse
-import org.knora.webapi.messages.v1storemessages.triplestoremessages.{SparqlSelectRequest, SparqlSelectResponse}
+import org.knora.webapi.messages.v1.store.triplestoremessages.{SparqlSelectRequest, SparqlSelectResponse}
 import org.knora.webapi.twirl.SearchCriterion
 import org.knora.webapi.util.ActorUtil._
 import org.knora.webapi.util.{DateUtilV1, FormatConstants}
