+++ webapi/src/main/scala/org/knora/webapi/responders/v1/UsersResponderV1.scala
@@ -24,10 +24,8 @@ import akka.actor.Status
 import akka.pattern._
 import org.knora.webapi._
 import org.knora.webapi.messages.v1.responder.projectmessages.{ProjectInfoByIRIGetRequest, ProjectInfoResponseV1, ProjectInfoType, ProjectInfoV1}
-import org.knora.webapi.messages.v1respondermessages.projectmessages.ProjectInfoType
-import org.knora.webapi.messages.v1respondermessages.triplestoremessages.SparqlSelectResponse
-import org.knora.webapi.messages.v1respondermessages.usermessages.{UserDataV1, UserProfileByUsernameGetRequestV1, UserProfileGetRequestV1, UserProfileV1, UsersResponderRequestV1}
-import org.knora.webapi.messages.v1storemessages.triplestoremessages.{SparqlSelectRequest, SparqlSelectResponse}
+import org.knora.webapi.messages.v1.responder.usermessages.{UserDataV1, UserProfileByUsernameGetRequestV1, UserProfileGetRequestV1, UserProfileV1}
+import org.knora.webapi.messages.v1.store.triplestoremessages.{SparqlSelectRequest, SparqlSelectResponse}
 import org.knora.webapi.util.ActorUtil._
 
 import scala.concurrent.Future
@@ -38,7 +36,7 @@ import scala.concurrent.Future
 class UsersResponderV1 extends ResponderV1 {
 
     /**
-      * Receives a message extending [[UsersResponderRequestV1]], and returns a message of type [[UserProfileV1]]
+      * Receives a message extending [[org.knora.webapi.messages.v1.responder.usermessages.UsersResponderRequestV1]], and returns a message of type [[UserProfileV1]]
       * [[Status.Failure]]. If a serious error occurs (i.e. an error that isn't the client's fault), this
       * method first returns `Failure` to the sender, then throws an exception.
       */
