+++ webapi/src/main/scala/org/knora/webapi/routing/v1/ProjectsRouteV1.scala
@@ -23,7 +23,6 @@ package org.knora.webapi.routing.v1
 import akka.actor.ActorSystem
 import akka.event.LoggingAdapter
 import org.knora.webapi.messages.v1.responder.projectmessages.{ProjectInfoByIRIGetRequest, ProjectInfoType, ProjectsGetRequestV1}
-import org.knora.webapi.messages.v1respondermessages.projectmessages.ProjectInfoType
 import org.knora.webapi.routing.{Authenticator, RouteUtilV1}
 import org.knora.webapi.util.InputValidation
 import org.knora.webapi.{BadRequestException, SettingsImpl}
