+++ webapi/src/main/scala/org/knora/webapi/routing/v1/ResourceTypesRouteV1.scala
@@ -23,8 +23,7 @@ package org.knora.webapi.routing.v1
 import akka.actor.ActorSystem
 import akka.event.LoggingAdapter
 import org.knora.webapi.messages.v1.responder.ontologymessages._
-import org.knora.webapi.messages.v1respondermessages.ontologymessages._
-import org.knora.webapi.messages.v1respondermessages.usermessages.UserProfileV1
+import org.knora.webapi.messages.v1.responder.usermessages.UserProfileV1
 import org.knora.webapi.routing.{Authenticator, RouteUtilV1}
 import org.knora.webapi.util.InputValidation
 import org.knora.webapi.{BadRequestException, SettingsImpl}
