+++ webapi/src/main/scala/org/knora/webapi/routing/v1/ValuesRouteV1.scala
@@ -26,11 +26,9 @@ import akka.actor.ActorSystem
 import akka.event.LoggingAdapter
 import org.knora.webapi.messages.v1.responder.sipimessages.{SipiResponderConversionFileRequestV1, SipiResponderConversionPathRequestV1}
 import org.knora.webapi.messages.v1.responder.valuemessages._
-import org.knora.webapi.messages.v1respondermessages.resourcemessages.CreateResourceApiRequestV1
-import org.knora.webapi.messages.v1respondermessages.sipimessages.SipiResponderConversionFileRequestV1
-import org.knora.webapi.messages.v1respondermessages.usermessages.UserProfileV1
-import ApiValueV1JsonProtocol._
-import org.knora.webapi.messages.v1respondermessages.valuemessages._
+import org.knora.webapi.messages.v1.responder.valuemessages.ApiValueV1JsonProtocol._
+import org.knora.webapi.messages.v1.responder.resourcemessages.CreateResourceApiRequestV1
+import org.knora.webapi.messages.v1.responder.usermessages.UserProfileV1
 import org.knora.webapi.routing.v1.ResourcesRouteV1._
 import org.knora.webapi.routing.{Authenticator, RouteUtilV1}
 import org.knora.webapi.util.{DateUtilV1, InputValidation}
