+++ webapi/src/main/scala/org/knora/webapi/util/InputValidation.scala
@@ -30,7 +30,6 @@ import org.joda.time.DateTime
 import org.joda.time.format.DateTimeFormat
 import org.knora.webapi._
 import org.knora.webapi.messages.v1.responder.valuemessages.{CreateRichtextV1, StandoffPositionV1}
-import org.knora.webapi.messages.v1respondermessages.valuemessages.StandoffPositionV1
 
 
 /**
