+++ webapi/src/main/scala/org/knora/webapi/util/MessageUtil.scala
@@ -3,8 +3,7 @@ package org.knora.webapi.util
 import org.apache.commons.lang3.StringEscapeUtils
 import org.knora.webapi._
 import org.knora.webapi.messages.v1.responder.valuemessages._
-import org.knora.webapi.messages.v1respondermessages.resourcemessages.{LiteralValueType, ResourceCreateValueObjectResponseV1, ResourceCreateValueResponseV1}
-import org.knora.webapi.messages.v1respondermessages.valuemessages._
+import org.knora.webapi.messages.v1.responder.resourcemessages.{LiteralValueType, ResourceCreateValueObjectResponseV1, ResourceCreateValueResponseV1}
 
 import scala.reflect.runtime.{universe => ru}
