+++ webapi/src/main/twirl/queries/sparql/v1/generateInsertStatementsForStandoffLinks.scala.txt
@@ -20,7 +20,7 @@
 
 @import org.knora.webapi._
 @import org.knora.webapi.twirl.SparqlTemplateLinkUpdate
-@import org.knora.webapi.messages.v1respondermessages.valuemessages._
+@import org.knora.webapi.messages.v1.responder.valuemessages._
 
 @**
  * Generates statements to be added to the INSERT clause of a SPARQL update to insert direct links and LinkValues
