+++ webapi/src/main/twirl/queries/sparql/v1/generateWhereStatementsForCreateValue.scala.txt
@@ -20,7 +20,7 @@
 
 @import org.knora.webapi._
 @import org.knora.webapi.twirl.SparqlTemplateLinkUpdate
-@import org.knora.webapi.messages.v1respondermessages.valuemessages._
+@import org.knora.webapi.messages.v1.responder.valuemessages._
 
 @**
  * Generates statements to be added to the WHERE clause of a SPARQL update when creating a new value of a resource
